--- apps/client/src/features/menu/RundownMenu.module.scss
@@ -1,4 +1,0 @@
-.headerButtons {
-  text-align: right;
-  padding: 1.5rem 1rem 0 1rem;
-}
