+++ apps/client/src/features/rundown/RundownWrapper.tsx
@@ -1,6 +1,7 @@
 import Empty from '../../common/components/state/Empty';
 import useRundown from '../../common/hooks-query/useRundown';
 
+import RundownHeader from './rundown-header/RundownHeader';
 import Rundown from './Rundown';
 
 import styles from './Rundown.module.scss';
@@ -10,6 +11,7 @@ export default function RundownWrapper() {
 
   return (
     <div className={styles.rundownWrapper}>
+      <RundownHeader />
       {status === 'success' && data ? <Rundown data={data} /> : <Empty text='Connecting to server' />}
     </div>
   );
