+++ apps/client/src/features/rundown/rundown-header/RundownHeader.module.scss
@@ -1,0 +1,12 @@
+.header {
+  padding-left: 2rem;
+  padding-right: 1rem;
+  
+  display: flex;
+  gap: 1rem;
+  align-items: center;
+}
+
+.header > :last-child {
+  margin-left: auto;
+}
