+++ apps/client/src/features/rundown/rundown-header/RundownMenu.tsx
@@ -6,8 +6,8 @@ import { IoTimerOutline } from '@react-icons/all-files/io5/IoTimerOutline';
 import { IoTrashOutline } from '@react-icons/all-files/io5/IoTrashOutline';
 import { SupportedEvent } from 'ontime-types';
 
-import { useEventAction } from '../../common/hooks/useEventAction';
-import { useEventSelection } from '../rundown/useEventSelection';
+import { useEventAction } from '../../../common/hooks/useEventAction';
+import { useEventSelection } from '../useEventSelection';
 
 const RundownMenu = ({ children }: { children: ReactNode }) => {
   const { clearSelectedEvents } = useEventSelection();
